moved Header.styles.tsx Header.tsx HeaderMenu.tsx to ui/components/layout/Header/
